+++ Assets/Scripts/Editor/GUI/PlayerSafePointEditor.cs
@@ -13,20 +13,20 @@ namespace MetroidvaniaEditor.SafePlayerPoints
 
         private BoxBoundsHandle _boundsHandle;
 
-        private SerializedProperty _triggerSize, _triggerOffset, _relativePoint;
+        private SerializedProperty _triggerSize, _triggerOffset, _position;
 
         private void OnEnable()
         {
             _code = target as PlayerSafePoint;
             _triggerSize = serializedObject.FindProperty("m_triggerSize");
-            _relativePoint = serializedObject.FindProperty("m_relativePoint");
+            _position = serializedObject.FindProperty("m_position");
             _triggerOffset = serializedObject.FindProperty("m_triggerOffset");
 
             _boundsHandle = new BoxBoundsHandle
             {
                 axes = PrimitiveBoundsHandle.Axes.X | PrimitiveBoundsHandle.Axes.Y
             };
-            _boundsHandle.SetColor(GizmosColor.instance.safePoints.area);
+            _boundsHandle.SetColor(GizmosColor.instance.safePoints.handles);
         }
 
         private void OnSceneGUI()
@@ -37,8 +37,8 @@ namespace MetroidvaniaEditor.SafePlayerPoints
             _triggerOffset.vector2Value = _boundsHandle.center - _code.transform.position;
             _triggerSize.vector2Value = _boundsHandle.size;
 
-            Handles.color = GizmosColor.instance.safePoints.area;
-            _relativePoint.vector2Value = Handles.FreeMoveHandle(_relativePoint.vector2Value, _code.transform.rotation, .25f, Vector3.one, Handles.DotHandleCap);
+            Handles.color = GizmosColor.instance.safePoints.handles;
+            _position.vector2Value = Handles.FreeMoveHandle(_position.vector2Value, Quaternion.identity, 0.25f, Vector3.zero, Handles.DotHandleCap);
 
             if (serializedObject.hasModifiedProperties)
                 serializedObject.ApplyModifiedProperties();