+++ Assets/Scripts/Units/Entities/SkeletonBehaviour.cs
@@ -26,14 +26,17 @@ namespace Metroidvania.Entities.Units
         [SerializeField, RangedValue(0, 1)] private RangedFloat m_RngStrength;
 
         [Header("Collisions")]
+        [SerializeField] private Collider2D m_Collider;
         [SerializeField] private Rect m_LeftLedge;
         [SerializeField] private Rect m_RightLedge;
+        [SerializeField] private Rect m_LeftHand;
+        [SerializeField] private Rect m_RightHand;
         [SerializeField] private LayerMask m_GroundLayer;
 
         [Header("Follow Target")]
         [SerializeField] private float m_FollowTargetSpeed;
-        [SerializeField] private float m_FollowTargetUpdateRate;
-        [SerializeField] private RangedFloat m_FollowTargetYOffset;
+        [UnityEngine.Serialization.FormerlySerializedAs("m_FollowTargetYOffset")]
+        [SerializeField] private RangedFloat m_FollowTargetVerticalView;
 
         [Header("Patrol")]
         [SerializeField, Range(0, 100)] private float m_PatrolChance;
@@ -65,9 +68,17 @@ namespace Metroidvania.Entities.Units
 
         private int facingDirection { get; set; }
         private float life { get; set; }
+        private float normalizedSpeedFactor { get; set; }
 
         private bool onLeftLedge { get; set; }
         private bool onRightLedge { get; set; }
+        private bool onLeftWall { get; set; }
+        private bool onRightWall { get; set; }
+
+        private bool leftObstructed => !onLeftLedge || onLeftWall;
+        private bool rightObstructed => !onRightLedge || onRightWall;
+
+        private bool targetObstructed { get; set; }
 
         private IdleState _idleState;
         private PatrolState _patrolState;
@@ -85,6 +96,7 @@ namespace Metroidvania.Entities.Units
 
             Flip(m_StartFacingRight ? 1 : -1);
             life = m_StartLife;
+            normalizedSpeedFactor = m_RngStrength.RandomRange();
 
             _idleState = new IdleState(this);
             _patrolState = new PatrolState(this);
@@ -96,12 +108,14 @@ namespace Metroidvania.Entities.Units
             SwitchState(_idleState);
         }
 
-        protected override void FixedUpdate()
+        private void FixedUpdate()
         {
             onLeftLedge = Physics2D.OverlapBox(rb.position + m_LeftLedge.center, m_LeftLedge.size, 0, m_GroundLayer);
             onRightLedge = Physics2D.OverlapBox(rb.position + m_RightLedge.center, m_RightLedge.size, 0, m_GroundLayer);
+            onLeftWall = Physics2D.OverlapBox(rb.position + m_LeftHand.center, m_LeftHand.size, 0, m_GroundLayer);
+            onRightWall = Physics2D.OverlapBox(rb.position + m_RightHand.center, m_RightHand.size, 0, m_GroundLayer);
 
-            base.FixedUpdate();
+            targetObstructed = targetFinder.hasFocusedTarget && targetFinder.IsObstructed(targetFinder.focusedTarget.position);
         }
 
         public void OnTakeHit(PlayerHitData hitData)
@@ -154,11 +168,24 @@ namespace Metroidvania.Entities.Units
             transform.localScale = scale;
         }
 
-        private bool ShouldFollowTargetAt(Vector2 targetPosition)
+        private void LookAtFocusedTarget()
         {
-            float direction = Mathf.Sign(targetPosition.x - rb.position.x);
-            return !((direction < 0 && !onLeftLedge) || (direction > 0 && !onRightLedge));
+            if (targetFinder.hasFocusedTarget)
+                FlipIfShould(targetFinder.focusedTarget.position.x);
         }
+
+        private bool ShouldFollowFocusedTarget() => !targetObstructed && targetFinder.hasFocusedTarget && CanWalkToPosition(targetFinder.focusedTarget.position);
+
+        private bool CanWalkToPosition(Vector2 targetPosition) => CanWalkToDirection(targetPosition.x - rb.position.x) && IsInsideVerticalView(targetPosition.y);
+
+        private bool CanWalkToDirection(float direction)
+        {
+            bool leftObsolete = direction < 0 && leftObstructed;
+            bool rightObsolete = direction > 0 && rightObstructed;
+            return !leftObsolete && !rightObsolete;
+        }
+
+        private bool IsInsideVerticalView(float position) => m_FollowTargetVerticalView.Contains(position - rb.position.y);
 
 #if UNITY_EDITOR
         private void OnValidate()
@@ -178,12 +205,14 @@ namespace Metroidvania.Entities.Units
             gizmos.SetColor(Color.red).DrawWireDisc(transform.position, m_DistanceToAttack);
 
             gizmos.SetColor(Color.blue)
-                .DrawWireDisc(transform.position + new Vector3(0, m_FollowTargetYOffset.min), 0.25f)
-                .DrawWireDisc(transform.position + new Vector3(0, m_FollowTargetYOffset.max), 0.25f);
+                .DrawWireDisc(transform.position + new Vector3(0, m_FollowTargetVerticalView.min), 0.25f)
+                .DrawWireDisc(transform.position + new Vector3(0, m_FollowTargetVerticalView.max), 0.25f);
 
             gizmos.SetColor(Color.green)
                 .DrawWireSquare(m_LeftLedge.center + (Vector2)transform.position, m_LeftLedge.size)
-                .DrawWireSquare(m_RightLedge.center + (Vector2)transform.position, m_RightLedge.size);
+                .DrawWireSquare(m_RightLedge.center + (Vector2)transform.position, m_RightLedge.size)
+                .DrawWireSquare(m_LeftHand.center + (Vector2)transform.position, m_LeftHand.size)
+                .DrawWireSquare(m_RightHand.center + (Vector2)transform.position, m_RightHand.size);
 
             void DrawAttack(AttackData attack)
             {
@@ -217,7 +246,7 @@ namespace Metroidvania.Entities.Units
 
             public override void LogicUpdate()
             {
-                if (entity.targetFinder.hasFocusedTarget && entity.ShouldFollowTargetAt(entity.targetFinder.focusedTarget.position))
+                if (entity.ShouldFollowFocusedTarget())
                 {
                     entity.SwitchState(entity._followTargetState);
                     return;
@@ -259,20 +288,16 @@ namespace Metroidvania.Entities.Units
                     return;
                 }
 
-                if (entity.targetFinder.hasFocusedTarget && entity.ShouldFollowTargetAt(entity.targetFinder.focusedTarget.position))
+                if (entity.ShouldFollowFocusedTarget())
                 {
                     entity.SwitchState(entity._followTargetState);
                     return;
                 }
 
-                int direction = entity.facingDirection;
-                if (direction == 1 && !entity.onRightLedge || direction == -1 && !entity.onLeftLedge)
-                {
-                    direction *= -1;
-                    entity.Flip(direction);
-                }
+                if (!entity.CanWalkToDirection(entity.facingDirection))
+                    entity.Flip(entity.facingDirection * -1);
 
-                entity.rb.velocity = new Vector2(direction * entity.m_PatrolMoveSpeed, entity.rb.velocity.y);
+                entity.rb.velocity = new Vector2(entity.facingDirection * entity.m_PatrolMoveSpeed * entity.normalizedSpeedFactor, entity.rb.velocity.y);
             }
         }
 
@@ -290,19 +315,6 @@ namespace Metroidvania.Entities.Units
 
             public override void LogicUpdate()
             {
-                if (!entity.targetFinder.hasFocusedTarget
-                    || !entity.m_FollowTargetYOffset.Contains(entity.targetFinder.focusedTarget.position.y - entity.rb.position.y))
-                {
-                    entity.SwitchState(entity._idleState);
-                    return;
-                }
-
-                if (entity.TryEnterAttackState())
-                    return;
-            }
-
-            public override void PhysicsUpdate()
-            {
                 Vector2 entityPosition = entity.rb.position;
                 Vector2 focusedTargetPosition = entity.targetFinder.focusedTarget.position;
 
@@ -310,13 +322,16 @@ namespace Metroidvania.Entities.Units
 
                 float direction = Mathf.Sign(focusedTargetPosition.x - entityPosition.x);
 
-                if (!entity.ShouldFollowTargetAt(focusedTargetPosition))
+                if (!entity.ShouldFollowFocusedTarget())
                 {
                     entity.SwitchState(entity._idleState);
                     return;
                 }
 
-                entity.rb.velocity = new Vector2(entity.m_FollowTargetSpeed * direction, entity.rb.velocity.y);
+                if (entity.TryEnterAttackState())
+                    return;
+
+                entity.rb.velocity = new Vector2(entity.m_FollowTargetSpeed * direction * entity.normalizedSpeedFactor, entity.rb.velocity.y);
             }
 
             public override void Exit()
@@ -343,7 +358,7 @@ namespace Metroidvania.Entities.Units
                 _performedFirstAttack = false;
                 _performedSecondAttack = false;
                 _elapsedTime = 0;
-                LookAtFocusedTarget();
+                entity.LookAtFocusedTarget();
                 entity.PlayAnimation("Attack");
             }
 
@@ -369,19 +384,18 @@ namespace Metroidvania.Entities.Units
 
             private void PerformAttack(AttackData attackData)
             {
-                LookAtFocusedTarget();
+                entity.LookAtFocusedTarget();
 
-                entity.rb.MovePosition(entity.rb.position + new Vector2(attackData.moveOffset * entity.facingDirection * entity.m_RngStrength.RandomRange(), 0));
+                float hitDistance = attackData.moveOffset * entity.m_RngStrength.RandomRange();
+                RaycastHit2D hit = Physics2D.Raycast(entity.rb.position, new Vector2(entity.facingDirection, 0).normalized, hitDistance, entity.m_GroundLayer);
+                if (hit)
+                    entity.rb.MovePosition(new Vector2(hit.point.x - (0.5f * entity.facingDirection * entity.m_Collider.bounds.size.x), hit.point.y));
+                else
+                    entity.rb.MovePosition(new Vector2(entity.rb.position.x + hitDistance * entity.facingDirection, entity.rb.position.y));
 
                 CombatUtility.CastEntityBoxHit(entity.rb.position + attackData.attackCollision.center * entity.transform.localScale,
                     attackData.attackCollision.size, _hits, entity.m_PlayerLayer, attackData.damage,
                     CombatUtility.FromFacingDirection(attackData.knockbackForce, entity.facingDirection));
-            }
-
-            private void LookAtFocusedTarget()
-            {
-                if (entity.targetFinder.hasFocusedTarget)
-                    entity.FlipIfShould(entity.targetFinder.focusedTarget.position.x);
             }
         }