+++ Assets/Scripts/Utility/CoroutinesUtility.cs
@@ -6,7 +6,7 @@ namespace Metroidvania
     /// <summary>Utility class for Unity Coroutines</summary>
     public static class CoroutinesUtility
     {
-        private static readonly Dictionary<float, WaitForSeconds> k_YieldSecondsWait = new();
+        private static readonly Dictionary<float, WaitForSeconds> k_YieldSecondsWait = new Dictionary<float, WaitForSeconds>();
 
         /// <summary>Use this method to get a cached <see cref="WaitForSeconds"/></summary>
         /// <param name="time">The yield time</param>
@@ -16,6 +16,12 @@ namespace Metroidvania
             if (!k_YieldSecondsWait.TryGetValue(time, out WaitForSeconds seconds))
                 k_YieldSecondsWait[time] = seconds = new WaitForSeconds(time);
             return seconds;
+        }
+
+        private static readonly WaitForEndOfFrame _waitForEndOfFrame = new WaitForEndOfFrame();
+        public static WaitForEndOfFrame GetWaitForEndOfFrame()
+        {
+            return _waitForEndOfFrame;
         }
 
         /// <summary>Stop the coroutine in behaviour and make its reference null</summary>